+++ packages/flyout/lib/src/flyout/directives/flyout.directive.ts
@@ -82,7 +82,9 @@ export class FlyoutDirective implements OnDestroy {
 			)
 			.subscribe((state: FlyoutState) => {
 				this.isOpened = state === FlyoutState.OPEN;
-				this.flyoutZone.isExpanded = this.isOpened;
+				if (this.flyoutZone) {
+					this.flyoutZone.isExpanded = this.isOpened;
+				}
 
 				if (this.isOpened) {
 					this.opened.emit();