+++ packages/forms/examples/src/forms/pages/upload/upload.page.ts
@@ -48,10 +48,14 @@ export class AppModule {};`;
 	url: 'api/upload',
 };`;
 
-	public uploadExampleHTML1 = `<aui-upload [options]="dropzone1" (selectUploadedFiles)="onUpload($event)">
+	public uploadExampleHTML1 = `<label class="a-input__label" for="my-upload">Upload files</label>
+<aui-upload
+	data-id="my-upload"
+	[options]="dropzone1"
+	(selectUploadedFiles)="onUpload($event)">
 	<div class="m-upload__message">
 		Drag your files here or click to upload
-  </div>
+	</div>
   <div class="m-upload__description">
     Optional description message
   </div>
@@ -65,7 +69,9 @@ export class AppModule {};`;
 	url: 'api/upload',
 };`;
 
-	public uploadExampleHTML2 = `<aui-upload [options]="dropzone2" (selectUploadedFiles)="onUpload($event)">
+	public uploadExampleHTML2 = `<aui-upload
+	[options]="dropzone2"
+	(selectUploadedFiles)="onUpload($event)">
 	<div class="m-upload__button">
 		Select your files to upload
   </div>
@@ -142,16 +148,16 @@ this.uploader.uploadFiles(this.queuedFiles).subscribe(
  * Using a custom upload service
  */
 
-// public uploadFiles(): void {
-//	if (!this.queuedFiles.length) {
-// 		return;
-// 	}
-// 	this.customService.postFile(this.queuedFiles).subscribe(res => {
-// 		this.result = res;
-// 	}, (error) => {
-// 		console.log(error);
-// 	});
-// }`;
+public uploadFiles(): void {
+	if (!this.queuedFiles.length) {
+		return;
+	}
+	this.customService.postFile(this.queuedFiles).subscribe(res => {
+		this.result = res;
+	}, (error) => {
+		console.log(error);
+	});
+}`;
 
 	public uploadExampleSCSS = `// don't display the default upload button
 ::ng-deep .a-upload-queue__wrapper {
@@ -161,6 +167,7 @@ this.uploader.uploadFiles(this.queuedFiles).subscribe(
 }`;
 
 	public uploadExampleHTML3 = `<aui-upload-zone
+	ariaId="my-other-upload"
 	[uploader]="uploader"
 	(queuedFiles)="onQueuedFiles($event)"
 	(uploadedFiles)="onUploadedFiles($event)"
@@ -184,7 +191,7 @@ this.uploader.uploadFiles(this.queuedFiles).subscribe(
 				type="button"
 				(click)="reloadErrors()"
 				class="m-upload__delete a-button-transparent a-button--danger a-button--small has-icon">
-				<span class="fa fa-close"></span>
+				<span class="u-screen-reader-only">Remove</span>
 			</button>
 		</li>
 	</ul>