+++ packages/forms/lib/src/upload/components/upload/upload.component.ts
@@ -9,15 +9,21 @@ import { Uploader } from '../../classes/uploader.class';
 	templateUrl: './upload.component.html',
 })
 export class UploadComponent implements OnInit {
+	@Input() public id = '';
+	@Input() public ariaLabelRemove = 'Verwijder';
 	@Input() public options: UploadOptions = UPLOAD_OPTIONS_DEFAULT;
 	@Output() public selectUploadedFiles: EventEmitter<Object[]> = new EventEmitter<Object[]>();
 
 	public uploader;
+	public ariaId;
 	public uploadedFiles: Object[] = [];
 	public invalidFiles: InvalidFile[] = [];
 	public queuedFiles: File[] = [];
 
 	ngOnInit() {
+		if (!this.id) {
+			this.ariaId = 'aui-upload-' + Math.random().toString(36).substring(2);
+		}
 		this.uploader = new Uploader(this.options);
 	}