+++ packages/forms/lib/src/upload/components/validation-list/validation-list.component.ts
@@ -10,7 +10,7 @@ import { ValidationMessagesService } from '../../services/validation-messages.se
 })
 export class ValidationListComponent {
 	@Input() public invalidFiles: InvalidFile[] = [];
-	@Input() public ariaLabelRemove = 'Verwijder upload';
+	@Input() public ariaLabelRemove = 'Verwijder';
 
 	constructor(private messagesService: ValidationMessagesService) {}